Day 1 challenge
This one was super easy. In the spirit of "for fun" coding, I re-coded part2 to use readline to read the file line-by-line instead of all at once. I also updated it to track just the top X elves in memory instead of storing all of the calories for all of the elves, sorting that, and then adding up the top 3. The readline turned out to be WAY WAY faster. Went from 3.18ms down to 0.835ms!